fix(asset-room): add colunm quanity

+++ src/main/java/com/example/datn_qlnt_manager/dto/projection/AssetRoomView.java
@@ -13,8 +13,9 @@ import lombok.experimental.FieldDefaults;
 public class AssetRoomView {
     String id;
     String roomCode;
-    int totalAssets;
+    long totalAssets;
     RoomType roomType;
     RoomStatus status;
     String description;
 }
+
